+++ src/main/java/kr/hhplus/be/server/domain/event/outbox/OutboxEventListener.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.domain.event.outbox;
+
+import kr.hhplus.be.server.domain.event.payment.PaymentEvent;
+
+public interface OutboxEventListener {
+
+    void paymentInitHandler(PaymentEvent event);
+}
