+++ src/main/java/kr/hhplus/be/server/domain/event/payment/PaymentEventListener.java
@@ -2,5 +2,7 @@ package kr.hhplus.be.server.domain.event.payment;
 
 public interface PaymentEventListener {
 
+    void sendMessageHandler(PaymentEvent event);
+
     void paymentSuccessHandle(PaymentSuccessEvent event);
 }
