+++ src/main/java/kr/hhplus/be/server/infra/repository/impl/OutboxRepositoryImpl.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.infra.repository.impl;
 
 import kr.hhplus.be.server.domain.entity.Outbox;
-import kr.hhplus.be.server.domain.event.OutboxEvent;
+import kr.hhplus.be.server.domain.event.outbox.OutboxEvent;
 import kr.hhplus.be.server.domain.repository.OutboxRepository;
 import kr.hhplus.be.server.infra.repository.jpa.OutboxJpaRepository;
 import lombok.RequiredArgsConstructor;
